# perf(*) reduce to minimal string concat per request