--- a/kong/plugins/prometheus/prometheus.lua
+++ b/kong/plugins/prometheus/prometheus.lua
@@ -137,14 +137,17 @@ local function check_metric_and_label_names(metric_name, label_names)
 end
 
 -- Makes a shallow copy of a table
-local function copy_table(table)
-  local new = {}
-  if table ~= nil then
-    for k, v in ipairs(table) do
-      new[k] = v
+local copy_table, err = pcall(require, "table.clone")
+if err then
+  copy_table = function(table)
+    local new = {}
+    if table ~= nil then
+      for k, v in ipairs(table) do
+        new[k] = v
+      end
     end
+    return new
   end
-  return new
 end
 
 -- Construct bucket format for a list of buckets.
@@ -241,7 +244,7 @@ local function inc(self, value, label_values)
     return
   end
 
-  local k, _, err
+  local k, err
   k, err = lookup_or_create(self, label_values)
   if err then
     self._log_error(err)
@@ -258,10 +261,7 @@ local function inc(self, value, label_values)
     end
     self._counter = c
   end
-  _, err = c:incr(k, value, 0)
-  if err then
-    self._log_error(err)
-  end
+  c:incr(k, value, 0)
 end
 
 local function del(self, label_values)
@@ -317,36 +317,23 @@ local function observe(self, value, label_values)
     end
     self._counter = c
   end
-  local _
   -- count
-  _, err = c:incr(keys[1], 1, 0)
-  if err then
-    self._log_error(err)
-  end
+  c:incr(keys[1], 1, 0)
 
   -- sum
-  _, err = c:incr(keys[2], value, 0)
-  if err then
-    self._log_error(err)
-  end
+  c:incr(keys[2], value, 0)
 
   local seen = false
   for i, bucket in ipairs(self.bucket) do
     if value <= bucket then
-      _, err = c:incr(keys[2+i], 1, 0)
-      if err then
-        self._log_error(err)
-      end
+      c:incr(keys[2+i], 1, 0)
       seen = true
     elseif seen then
       break
     end
   end
   -- inf
-  _, err = c:incr(keys[self.bucket_count+3], 1, 0)
-  if err then
-    self._log_error(err)
-  end
+  c:incr(keys[self.bucket_count+3], 1, 0)
 end
 
 local function reset(self)