QPS on packet.net baremetal 4c8t E3-1578L v5 @ 2.00GHz
vanilla kong: 44917.29
master: 7586.99
this branch: ~30018.47~ 37244.20 (after each worker is warmed up with cache)

- move key search to a table based search, expensive string operations now only occur once per label group per worker. this is basically trading cpu with memory and requires measurement of memory consumption on high cardinality.
- lua-resty-counter is used by default
- reduced function call in hot path