@fffonion 2020-03-12
<img width="1202" alt="Screen Shot 2020-03-12 at 8 18 42 PM" src="https://user-images.githubusercontent.com/2135445/76520869-b1decb00-649e-11ea-8364-178bcc3a52e9.png">

exporter.lua now appears only less than 10% in stacktraces