@fffonion 2020-03-12
<img width="1229" alt="Screen Shot 2020-03-12 at 8 28 42 PM" src="https://user-images.githubusercontent.com/2135445/76521599-18b0b400-64a0-11ea-9241-7f68e3732bbc.png">

exporter.lua now appears only less than 10% in stacktraces
[1.svg.gz](https://github.com/Kong/kong-plugin-prometheus/files/4324137/1.svg.gz)