@fffonion 2020-03-12
scrape lock analysis:

without scrape:
```
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   269.84us  156.67us   3.16ms   91.81%
    Req/Sec    19.36k     2.04k   22.84k    75.12%
  387239 requests in 10.10s, 69.78MB read
Requests/sec:  38342.68
Transfer/sec:      6.91MB
```

with scrape every one second
```
Running 10s test @ http://10.11.9.129:8000
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     0.93ms    7.12ms 134.74ms   98.91%
    Req/Sec    18.99k     2.61k   23.10k    79.21%
  381683 requests in 10.10s, 68.78MB read
Requests/sec:  37790.38
Transfer/sec:      6.81MB
```

there's *some* impact especially with p95 latencies, but the overall QPS doesn't drop a lot