@hbagdi 2020-03-25
@fffonion Did you do any measurement on the amount of memory consumption or how it will increase as more routes/services are added?