@fffonion 2020-03-26
Yeah. A three-label metrics takes around 250 ~ 400 bytes for a single worker.

For example, in a 4 worker cluster:
```
status{"label1"="value1","label2"="value2","label1"="value3"} 1
status{"label1"="value4","label2"="value5","label1"="value6"} 2
```
will take 3200 bytes in total.

```
status{"label1"="value1","label2"="value2","label1"="value3"} 1
status{"label1"="value1","label2"="value2","label1"="value4"} 2
```
This will takes less if labels diverges at last. But the scale is still like ~200 ish per metrics.

[memory.lua.gz](https://github.com/Kong/kong-plugin-prometheus/files/4385624/memory.lua.gz)