@hbagdi 2020-04-20
```diff
diff -Naur kong/plugins/prometheus/prometheus.lua ~/play/nginx-lua-prometheus/prometheus.lua
--- kong/plugins/prometheus/prometheus.lua      2020-04-20 15:14:35.428147739 -0700
+++ /home/hbagdi/play/nginx-lua-prometheus/prometheus.lua       2020-04-20 15:11:57.085112212 -0700
@@ -678,4 +678,4 @@
     "Error while setting '", key, "' to '", value, "': '", err, "'")
 end
 
-return Prometheus
\ No newline at end of file
+return Prometheus

```

@fffonion There is a small diff between the vendored `prometheus.lua` file and the one in the master branch on the upstream. Can you fix that?
This is a huge improvement and this PR makes me very happy! Not just from the performance point of view but also seeing that we are completely aligned with the upstream module.